Updated dlls and functionalities, updated tests.
Now VOCASY system will specific a list of receivers to which send data,
this list will depend on the currently added handlers on the workflow
and on the received IsMutedMessages. Reworked classes to fit with such
change. Some tests were broken after the change, there will be an overall
fix to most of the tests in the next commit.